Extract the service layer into services.py (R1)
`__init__.py` had grown to ~1140 lines, most of it a single ~560-line
`_register_services` closure holding every service handler, plus the schemas and a
hand-maintained `_SERVICES` teardown tuple that had to be kept in sync with the
registration calls by hand.

Move the whole service surface into a new `services.py`:

- Schemas live alongside the handlers they validate.
- Handlers become module-level `async def _handle_*(hass, call)` functions, bound
  to `hass` at registration via a typed closure (`ServiceCall` doesn't carry `hass`
  on every supported HA version, so a partial/closure is required).
- A frozen `_ServiceDef` table is the single source of truth for both
  `async_register_services` and `async_unregister_services`, so registration and
  teardown can no longer drift.
- A `_translate_errors` decorator folds the boilerplate that every mutating handler
  repeated: `TaskValidationError -> invalid_task`, `AssetValidationError ->
  invalid_asset`, and `KeyError -> {task,asset}_not_found` (the two-placeholder
  `unknown_part` case stays explicit). Translation keys remain literal constants so
  the exception-translations AST drift guard still validates them.

`__init__.py` now holds only entry lifecycle (~176 lines). No behavior change:
same 26 services, same schemas, same response kinds, same localized errors.
